fix(prospect): unify themes, update imports + use red #E50000 for alert (#1551)
