--- packages/canopee-css/src/prospect-client/Form/ItemMessage/ItemMessageApollo.css
@@ -1,17 +1,0 @@
-@import "./ItemMessageCommon.css";
-
-@layer canopee {
-  .af-item-message {
-    --item-message-color: var(--red-alert-1000);
-
-    font-size: clamp(14px, 1.4vw, 16px);
-  }
-
-  .af-item-message--success {
-    --item-message-color: var(--green-1000);
-  }
-
-  .af-item-message--warning {
-    --item-message-color: var(--orange-1000);
-  }
-}
