--- packages/canopee-css/src/prospect-client/Form/ItemMessage/ItemMessageLF.css
@@ -1,15 +1,0 @@
-@import "./ItemMessageCommon.css";
-
-@layer canopee {
-  .af-item-message {
-    --item-message-color: var(--red-alert-1200);
-  }
-
-  .af-item-message--success {
-    --item-message-color: var(--green-1000);
-  }
-
-  .af-item-message--warning {
-    --item-message-color: var(--orange-1000);
-  }
-}
