+++ packages/canopee-react/src/prospect-client/Form/Checkbox/CheckboxText/__tests__/CheckboxText.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { ItemMessage } from "../../../ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../../../ItemMessage/ItemMessage";
 import { Checkbox } from "../../Checkbox/CheckboxCommon";
 import {
   CheckboxTextCommon,
