+++ packages/canopee-react/src/prospect-client/Form/Dropdown/DropdownLF.tsx
@@ -1,6 +1,6 @@
 import "@axa-fr/canopee-css/client/Form/Dropdown/DropdownLF.css";
 import { ItemLabel } from "../ItemLabel/ItemLabelLF";
-import { ItemMessage } from "../ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../ItemMessage/ItemMessage";
 import { DropdownCommon, type DropdownProps } from "./DropdownCommon";
 
 export const Dropdown = (props: DropdownProps) => (
