+++ packages/canopee-react/src/prospect-client/Form/FileUpload/InputFile/__tests__/InputFile.test.tsx
@@ -11,7 +11,7 @@ import {
   ItemLabelCommon,
   type ItemLabelProps,
 } from "../../../ItemLabel/ItemLabelCommon";
-import { ItemMessage } from "../../../ItemMessage/ItemMessageApollo";
+import { ItemMessage } from "../../../ItemMessage/ItemMessage";
 import { InputFileCommon, type InputFileProps } from "../InputFileCommon";
 
 const ButtonMock = (props: ButtonProps) => (
