+++ packages/canopee-react/src/prospect-client/Form/FileUpload/ItemFile/ItemFileLF.tsx
@@ -3,7 +3,7 @@ import visibilityIcon from "@material-symbols/svg-400/outlined/visibility-fill.s
 import { ClickIcon } from "../../../ClickIcon/ClickIconLF";
 import { Icon } from "../../../Icon/IconLF";
 import { Spinner } from "../../../Spinner/SpinnerLF";
-import { ItemMessage } from "../../ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../../ItemMessage/ItemMessage";
 import { ItemFileCommon, type ItemFileProps } from "./ItemFileCommon";
 import "@axa-fr/canopee-css/client/ContentItemMono/ContentItemMonoLF.css";
 import "@axa-fr/canopee-css/client/Form/FileUpload/ItemFile/ItemFileLF.css";
