+++ packages/canopee-react/src/prospect-client/Form/InputDate/InputDateApollo.tsx
@@ -1,6 +1,6 @@
 import "@axa-fr/canopee-css/prospect/Form/InputDate/InputDateApollo.css";
 import { ItemLabel } from "../ItemLabel/ItemLabelApollo";
-import { ItemMessage } from "../ItemMessage/ItemMessageApollo";
+import { ItemMessage } from "../ItemMessage/ItemMessage";
 import { InputDateCommon, type InputDateProps } from "./InputDateCommon";
 
 export const InputDate = ({ ...props }: InputDateProps) => (
