+++ packages/canopee-react/src/prospect-client/Form/InputPhone/InputPhoneApollo.tsx
@@ -1,6 +1,6 @@
 import "@axa-fr/canopee-css/prospect/Form/InputPhone/InputPhoneApollo.css";
 import { ItemLabel } from "../ItemLabel/ItemLabelApollo";
-import { ItemMessage } from "../ItemMessage/ItemMessageApollo";
+import { ItemMessage } from "../ItemMessage/ItemMessage";
 import { InputPhoneCommon, type InputPhoneProps } from "./InputPhoneCommon";
 import { InputTextAtom } from "../InputTextAtom/InputTextAtomApollo";
 import { Icon } from "../../Icon/IconApollo";
