+++ packages/canopee-react/src/prospect-client/Form/InputText/InputTextApollo.tsx
@@ -1,6 +1,6 @@
 import "@axa-fr/canopee-css/prospect/Form/InputText/InputTextApollo.css";
 import { ItemLabel } from "../ItemLabel/ItemLabelApollo";
-import { ItemMessage } from "../ItemMessage/ItemMessageApollo";
+import { ItemMessage } from "../ItemMessage/ItemMessage";
 import { InputTextCommon, type InputTextProps } from "./InputTextCommon";
 import { InputTextAtom } from "../InputTextAtom/InputTextAtomApollo";
