+++ packages/canopee-react/src/prospect-client/Form/InputText/InputTextLF.tsx
@@ -1,6 +1,6 @@
 import "@axa-fr/canopee-css/client/Form/InputText/InputTextLF.css";
 import { ItemLabel } from "../ItemLabel/ItemLabelLF";
-import { ItemMessage } from "../ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../ItemMessage/ItemMessage";
 import { InputTextCommon, type InputTextProps } from "./InputTextCommon";
 import { InputTextAtom } from "../InputTextAtom/InputTextAtomLF";
