+++ packages/canopee-react/src/prospect-client/Form/ItemMessage/ItemMessage.tsx
@@ -1,0 +1,53 @@
+import "@axa-fr/canopee-css/prospect/Form/ItemMessage/ItemMessageAll.css";
+import successIcon from "@material-symbols/svg-400/outlined/check_circle-fill.svg";
+import errorIcon from "@material-symbols/svg-400/outlined/error-fill.svg";
+import warningIcon from "@material-symbols/svg-400/outlined/warning-fill.svg";
+import type { ReactNode } from "react";
+import { Icon } from "../../Icon/IconCommon";
+
+export const itemMessageVariants = {
+  error: "error",
+  success: "success",
+  warning: "warning",
+} as const;
+
+export type ItemMessageVariants = keyof typeof itemMessageVariants;
+
+export type ItemMessageProps = {
+  message?: ReactNode;
+  id?: string;
+  messageType?: ItemMessageVariants;
+};
+
+const getIcon = (messageType: ItemMessageVariants) => {
+  if (messageType === "success") return successIcon;
+  if (messageType === "warning") return warningIcon;
+  return errorIcon;
+};
+
+export const ItemMessage = ({
+  message,
+  id,
+  messageType = "error",
+}: ItemMessageProps) => {
+  if (!message) {
+    return null;
+  }
+
+  return (
+    <small
+      id={id}
+      className={`af-item-message af-item-message--${messageType}`}
+      role={messageType === "success" ? undefined : "alert"}
+      aria-live="assertive"
+    >
+      <Icon
+        src={getIcon(messageType)}
+        size="XS"
+        variant={messageType}
+        aria-hidden="true"
+      />
+      <span className="af-item-message__message">{message}</span>
+    </small>
+  );
+};
