+++ packages/canopee-react/src/prospect-client/Form/Radio/CardRadioGroup/CardRadioGroupLF.tsx
@@ -1,4 +1,4 @@
-import { ItemMessage } from "../../ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../../ItemMessage/ItemMessage";
 import { CardRadio } from "../CardRadio/CardRadioLF";
 import {
   CardRadioGroupCommon,
