+++ packages/canopee-react/src/prospect-client/Form/Radio/CardRadioGroup/__tests__/CardRadioGroup.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { axe } from "jest-axe";
 import { describe, expect, it } from "vitest";
-import { ItemMessage } from "../../../ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../../../ItemMessage/ItemMessage";
 import { CardRadio } from "../../CardRadio/CardRadioLF";
 import {
   CardRadioGroupCommon,
