+++ packages/canopee-react/src/prospect-client/List/ContentItemDuo/ContentItemDuoLF.tsx
@@ -4,7 +4,7 @@ import {
   ContentItemDuoCommon,
   type ContentItemDuoProps,
 } from "./ContentItemDuoCommon";
-import { ItemMessage } from "../../Form/ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../../Form/ItemMessage/ItemMessage";
 
 export const ContentItemDuo = (props: ContentItemDuoProps) => (
   <ContentItemDuoCommon
