+++ packages/canopee-react/src/prospect-client/Stepper/StepperLF.tsx
@@ -1,5 +1,5 @@
 import "@axa-fr/canopee-css/client/Stepper/StepperLF.css";
-import { ItemMessage } from "../Form/ItemMessage/ItemMessageLF";
+import { ItemMessage } from "../Form/ItemMessage/ItemMessage";
 import { ProgressBarGroup } from "../ProgressBarGroup/ProgressBarGroupLF";
 import { StepperCommon, type StepperProps } from "./StepperCommon";
