+++ packages/canopee-react/src/prospect.ts
@@ -63,7 +63,7 @@ export {
   ItemMessage,
   itemMessageVariants,
   type ItemMessageVariants,
-} from "./prospect-client/Form/ItemMessage/ItemMessageApollo";
+} from "./prospect-client/Form/ItemMessage/ItemMessage";
 export {
   ItemMultiSelect,
   type ItemMultiSelectProps,
